VIPER changes
