+++ Mobile-Observatory/Modules/ArticleModule/Presenter/ArticleDetailPresenter.swift
