+++ Mobile-Observatory/Modules/ArticleModule/View/ArticleTableViewCell.swift
